// data/foodb-models/db.d.ts
export interface foodcomexCompoundsAttribute {
  id:number;
  compoundId:number;
  origin?:string;
  storageForm?:string;
  maximumQuantity?:string;
  storageCondition?:string;
  contactName?:string;
  contactAddress?:string;
  contactEmail?:string;
  createdAt:Date;
  updatedAt:Date;
  export?:number;
  purity?:string;
  description?:string;
  spectraDetails?:string;
  deliveryTime?:string;
  stability?:string;
  adminUserId?:number;
  publicId:string;
  casNumber?:string;
  taxonomyClass?:string;
  taxonomyFamily?:string;
  experimentalLogp?:string;
  experimentalSolubility?:string;
  meltingPoint?:string;
  foodOfOrigin?:string;
  productionMethodReferenceText?:string;
  productionMethodReferenceFileName?:string;
  productionMethodReferenceContentType?:string;
  productionMethodReferenceFileSize?:number;
  productionMethodReferenceUpdatedAt?:Date;
  elementalFormula?:string;
  minimumQuantity?:string;
  quantityUnits?:string;
  availableSpectra?:string;
  storageConditions?:string;
}
export interface foodcomexCompoundsInstance extends Sequelize.Instance<foodcomexCompoundsAttribute>, foodcomexCompoundsAttribute { }